+++ src/lr-wpan/model/lr-wpan-fields.h
@@ -144,7 +144,6 @@ public:
    */
   Buffer::Iterator Deserialize (Buffer::Iterator i);
 
-
 private:
   // Superframe Specification field
   // See IEEE 802.14.15-2011 5.2.2.1.2
@@ -228,11 +227,11 @@ private:
 
   //GTS specification field
   uint8_t m_gtsSpecDescCount;            //!< GTS specification field Descriptor Count (Bit 0-2)
-                                         // GTS specification field Reserved (Not necessary) (Bit 3-6)
+  // GTS specification field Reserved (Not necessary) (Bit 3-6)
   uint8_t m_gtsSpecPermit;               //!< GTS specification field GTS Permit (Bit 7)
   //GTS Direction field
   uint8_t m_gtsDirMask;                  //!< GTS Direction field Directions Mask (Bit 0-6)
-                                         // GTS Direction field Reserved (Not Necessary) (Bit 7)
+  // GTS Direction field Reserved (Not Necessary) (Bit 7)
   //GTS List
   gtsDescriptor m_gtsList[7];            //!< GTS List field (maximum descriptors stored == 7)
 };
@@ -341,6 +340,98 @@ private:
  * \returns The reference to the output stream.
  */
 std::ostream &operator << (std::ostream &os, const PendingAddrFields &pendingAddrFields);
+
+
+/**
+ * \ingroup lr-wpan
+ *
+ * Represent the Capability Information Field.
+ * See IEEE 802.15.4-2011   Section 5.3.1.2 Figure 50
+ */
+class CapabilityField
+{
+
+public:
+  CapabilityField ();
+  /**
+   * Get the size of the serialized Capability Information Field.
+   * \return the size of the serialized field.
+   */
+  uint32_t GetSerializedSize (void) const;
+  /**
+   * Serialize the entire Capability Information Field.
+   * \param i an iterator which points to where the Capability information field should be written.
+   * \return an iterator.
+   */
+  Buffer::Iterator Serialize (Buffer::Iterator i) const;
+  /**
+   * Deserialize the entire Capability Information Field.
+   * \param i an iterator which points to where the Capability information field should be read.
+   * \return an iterator.
+   */
+  Buffer::Iterator Deserialize (Buffer::Iterator i);
+  /**
+   * True if the device type is a Full Functional Device (FFD) false if is a Reduced Functional Device (RFD).
+   */
+  bool IsDeviceTypeFfd (void) const;
+  /**
+   * True if the device is receiving power from alternating current mains.
+   * \return True if the device is receiving power from alternating current mains.
+   */
+  bool IsPowSrcAvailable (void) const;
+  /**
+   * True if the device does not disable its receiver to conserve power during idle periods.
+   * \return True if the device does not disable its receiver to conserve power during idle periods.
+   */
+  bool IsReceiverOnWhenIdle (void) const;
+  /**
+   * True if the device is capable of sending and receiving cryptographically protected MAC frames.
+   * \return True if the device is capable of sending and receiving cryptographically protected MAC frames.
+   */
+  bool IsSecurityCapability (void) const;
+  /**
+   * True if the device wishes the coordinator to allocate a short address as result of the association procedure.
+   * \return True if the device wishes the coordinator to allocate a short address as result of the association procedure.
+   */
+  bool IsShortAddrAllocOn (void) const;
+  /**
+   * Set the Device type in the Capability Information Field.
+   * True = full functional device (FFD)  False = reduced functional device (RFD).
+   * \param devType The device type described in the Capability Information Field.
+   */
+  void SetFfdDevice (bool devType);
+  /**
+   * Set the Power Source available flag in the Capability Information Field.
+   * \param pow Set true if a Power Source is available in the Capability Information Field.
+   */
+  void SetPowSrcAvailable (bool pow);
+  /**
+   * Indicate if the receiver is On on Idle
+   * \param rxIdle Set true if the receiver is on when Idle
+   */
+  void SetRxOnWhenIdle (bool rxIdle);
+  /**
+   * Set the Security Capability flag in the Capability Information Field.
+   * \param sec Set true if the device have Security Capabilities.
+   */
+  void SetSecurityCap (bool sec);
+  /**
+   * Set the Short Address Flag in the Capability Information Field.
+   * \param addrAlloc Describes whether or not the coordinator should allocate a short
+   *                  address in the association process.
+   */
+  void SetShortAddrAllocOn (bool addrAlloc);
+
+private:
+
+  bool m_deviceType;          //!< Capability Information Field, Device Type  (bit 1)
+  bool m_powerSource;         //!< Capability Information Field, Power Source (bit 2)
+  bool m_receiverOnWhenIdle;  //!< Capability Information Field, Receiver On When Idle (bit 3)
+  bool m_securityCap;         //!< Capability Information Field, Security Capability (bit 6)
+  bool m_allocAddr;           //!< Capability Information Field, Allocate Address (bit 7)
+
+};
+std::ostream &operator << (std::ostream &os, const CapabilityField &capabilityField);
 
 
 }  //end namespace ns3