+++ CHANGES.md
@@ -22,7 +22,8 @@ Changes from ns-3.36 to ns-3.37
 
 ### Changes to existing API
 
- * Adds support for channel paging to the **LrWpanPhy** (only placeholder,  a single modulation/band is currently supported).
+* Adds support for channel paging to the **LrWpanPhy** (only placeholder,  a single modulation/band is currently supported).
+* Adds supporting structures used by **LrWpanMac** (PAN descriptor, Command Payload Header, Capability Field).
 
 ### Changes to build system